Move the model code back to the main app for now
Too much cross-pollution from specs, etc. to have engine models currently.
