+++ app/models/file_downloaders/basic_authentication.rb
@@ -1,0 +1,16 @@
+# frozen_string_literal: true
+
+module FileDownloaders
+  class BasicAuthentication < BaseDownloader
+    attribute :username, :string
+    attribute :password, :string
+
+    with_options presence: true do
+      validates :username, :password
+    end
+
+    def authenticate(connection)
+      connection.request :basic_auth, username, password
+    end
+  end
+end
