--- engines/bops_api/app/models/bops_api/application_record.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-module BopsApi
-  class ApplicationRecord < ActiveRecord::Base
-    self.abstract_class = true
-  end
-end
