--- engines/bops_api/app/models/bops_api/file_downloaders/basic_authentication.rb
@@ -1,18 +1,0 @@
-# frozen_string_literal: true
-
-module BopsApi
-  module FileDownloaders
-    class BasicAuthentication < BaseDownloader
-      attribute :username, :string
-      attribute :password, :string
-
-      with_options presence: true do
-        validates :username, :password
-      end
-
-      def authenticate(connection)
-        connection.request :basic_auth, username, password
-      end
-    end
-  end
-end
