--- engines/bops_api/app/models/bops_api/file_downloaders/bearer_authentication.rb
@@ -1,17 +1,0 @@
-# frozen_string_literal: true
-
-module BopsApi
-  module FileDownloaders
-    class BearerAuthentication < BaseDownloader
-      attribute :token, :string
-
-      with_options presence: true do
-        validates :token
-      end
-
-      def authenticate(connection)
-        connection.request :authorization, "Bearer", token
-      end
-    end
-  end
-end
