--- engines/bops_api/app/models/bops_api/file_downloaders/header_authentication.rb
@@ -1,18 +1,0 @@
-# frozen_string_literal: true
-
-module BopsApi
-  module FileDownloaders
-    class HeaderAuthentication < BaseDownloader
-      attribute :key, :string
-      attribute :value, :string
-
-      with_options presence: true do
-        validates :key, :value
-      end
-
-      def authenticate(connection)
-        connection.headers[key] = value
-      end
-    end
-  end
-end
