--- engines/bops_api/app/models/bops_api/local_authority.rb
@@ -1,8 +1,0 @@
-# frozen_string_literal: true
-
-module BopsApi
-  class LocalAuthority < ApplicationRecord
-    has_many :users
-    has_many :planning_applications
-  end
-end
