+++ engines/bops_api/lib/bops_api/errors.rb
@@ -5,7 +5,6 @@ module BopsApi
     class BaseError < ::StandardError; end
     class InvalidRequestError < BaseError; end
     class InvalidSchemaError < BaseError; end
-    class MissingFileDownloaderType < BaseError; end
     class NotPermittedError < BaseError; end
     class SchemaNotFoundError < BaseError; end
   end
