+++ engines/bops_api/spec/requests/v2/planning_applications_spec.rb
@@ -96,7 +96,7 @@ RSpec.describe "BOPS API" do
       response "404", "when a local authority isn't found" do
         before do
           exception = ActiveRecord::RecordNotFound.new("Local authority not found")
-          allow(BopsApi::LocalAuthority).to receive(:find_by!).and_raise(exception)
+          allow(LocalAuthority).to receive(:find_by!).and_raise(exception)
         end
 
         schema "$ref" => "#/components/schemas/NotFoundError"
